back-end: chore: add and set up sequelize with data migrations
Add and configure Sequelize ORM for connecting with MariaDB. The initial setup includes Umzug migration tool which integrates Sequelize as a tool for interacting with MariaDB. The migrations are configured in the migrate.js script in the root of the project. The model/mysql includes two fundamental data models for the application. I've tried using the default Sequelize migration tool, but it does not allow flexible configuration. Umzug is a recommended tool by Sequelize team for this purpose.