back-end: chore: set up sequelize
Set up an instance of sequelize that includes sequelize models. To do that, I had to do a bit of gymnastics for the architecture to work. Due to my lack of experience with Sequelize, I thought I could define models and pass a database connection into the model. It turns out, that sequelize is the database connection and defines the models. The sequelize instance is being passed down from the server's entry point down to business logic in core functions so I should be able to use sequelize-mock in unit tests. But a real test is actaully doing it. This commit includes a simple sequelize model 'language'. model/mysql needs a clean up and a setup for executing raw queries that Sequelize can't handle nicely.